requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

# Database Connectivity
sqlalchemy>=1.4.0
mysql-connector-python>=8.0.0

# Environment Management
python-dotenv>=0.19.0

# Machine Learning
joblib>=1.1.0

# Data Visualization (for Jupyter notebooks)
matplotlib>=3.5.0
seaborn>=0.11.0

# Jupyter Notebook Support
jupyter>=1.0.0